+++ src/middlewares/middlewares.js
@@ -6,6 +6,7 @@ import { validateInvite } from './validateInvite';
 import { validateInviteId, validateUserId, validateUpvoteInput } from './validateUUID';
 import { validateAdmin } from './validateAdmin';
 import { validateUserById } from './validateUser';
+import { validateUUID } from './validateUUID';
 
 export {
   validateSigninFormData,
